x86sim: shrq shift count must be double (cc18_unsigned sim 7->6)
The shrq branch computed u / 2^c with c = mod(sim_opval(a), 64), and a
register-count operand (%cl) reads back as a RAW int64. Real MATLAB
promotes double/int64 to double (5/2 = 2.5, floor = 2); both MATLAB
clones evaluate integer arithmetic with round-half semantics (2.5 -> 3),
so unsigned register-count shifts of odd values ran one high. Only odd
operands divide inexactly, which is why cc18_unsigned2/ushr passed while
cc18_unsigned (a>>1, a=5) returned 7 vs gcc's 6. Force the count double
so the float path stays double on every runtime; behavior-neutral on
real MATLAB. Verified on matlab_in_rust + v1.3.47: cc18_unsigned 7->6,
cc18_unsigned2 111, cc18_ushr 0, cc18_ucmp 1 unchanged.

Also record the resolution in the divergences doc (A1/A2 were pre-fixed
in-repo; C's state hypothesis falsified - peephole_pass.m has no globals).

+++ src/x86sim.m
@@ -360,7 +360,7 @@ elseif m == 39           % sarq (arithmetic)
     c = mod(sim_opval(a), 64);
     sim_opstore(b, bitshift(sim_opval(b), -c, 'int64'), 64);
 elseif m == 44           % shrq (logical, unsigned)
-    c = mod(sim_opval(a), 64);
+    c = double(mod(sim_opval(a), 64));
     u = mod(double(sim_opval(b)), 18446744073709551616);
     sim_opstore(b, int64(floor(u / 2^c)), 64);
 elseif m == 18           % cmpq
